refactor: use double instead of float

+++ src/cache.h
@@ -283,11 +283,11 @@ class cache {
 			return _n_l1_hits;
 		}
 
-		float get_hit_ratio() {
+		double get_hit_ratio() {
 			if (_n_reqs == 0)
 				return 0;
 
-			return ((_n_hits + _n_l1_hits) / static_cast<float>(_n_reqs));
+			return ((_n_hits + _n_l1_hits) / static_cast<double>(_n_reqs));
 		}
 #endif /* __PROFILE__ */
